major paradigm change -- parse but do not load HDF5 file
HDF5 files can by *very* large, and having emacs load the file into a buffer
before clearing the buffer and parsing the file with python is wasted time.  The
gist of this change is that the file is no longer fully read into any buffer,
making the viewing and exploring of HDF5 files much faster. The changes here do
the following:

1. Define advice to go :before-until `find-file`
   - check the first 8 bytes of the file.  If it has the HDF5 signature, proceed
   - otherwise drop into `find-file`
   - use the buffer name "*hdf5: FILENAME*" rather than FILENAME (see rationale
   for item 2 below).

2. Define a new global variable `hdf5--buffer-filename` which is set in the
advising function and then is used in `hdf5-mode` to set `hdf5-mode-file`.   I
thought using `buffer-file-name` was a bit risky, since one could potentially
overwite the HDF5 file with the contents of the hdf5-mode buffer.

3. Remove .h5 and .hdf5 from auto-mode-alist, as they are not needed to trigger
`hdf5-mode`

4. Updated the dataset buffer name to work well with the new hdf5-mode buffer name.