(major change) Use root path to place cursor when navigating back
This commit removes `hdf5--backward-point-list' in favor of a variable that
keeps track of the last group in the root path BEFORE navigating backwards.
This allows for proper cursor placement after a jump-navigation, which the
--backward-point-list cannot determine.

This uses a reverse search for the string "<parent group>/" starting
from (point-max), so it could fail if that string shows up elsewhere in the
buffer.