ensure that the path of the CLI parse command is correct
This is useful in debugging, but it will also prevent the package from breaking
on upgrades.

`hdf-viewer-parse-command` is initialized to the copy of `h5parse.py` in the same
directory as `hdf5-viewer.el`, but if the path changes, like on an upgrade or if
one has multiple copies of the package and reloads the package, the path will
not update without the call to `makunbound`.