add global minor mode to allow user to enable/disable this package
New minor mode is `hdf5-viewer-find-file-mode` as suggested by @riscy (MELPA).

`hdf5-viewer-maybe-startup` has been renamed to `hdf5-viewer-bypass-find-file`,
as that better describes the action of the `find-file` advice.

`hdf5-viewer-mode` has been altered to issue a message to the user and exit if
`hdf5-viewer-find-file-mode` is not enabled.  I think this is necessary because
of the precautions taken in `...-bypass-find-file` to protect the raw data file
from corruption.